refactor: page 객체를 감싸는 실수 해결
